Prevent mention spamming
I am worried that someone could get bot to mention other people and direct spam/abuse at them.

Mastodon uses a regex that expects an '@' symbol to exist for something to count as a mention. As far as I can see there isn't a way to escape mentions so instead I'm replacing them with a similar-but-different unicode character - the 'SMALL COMMERCIAL AT'

[1] https://github.com/mastodon/mastodon/blob/f682478/app/models/account.rb#LL66C3-L66C13